chore: Format code using forge fmt

+++ src/week1/BondingCurve.sol
@@ -19,20 +19,17 @@ error CooldownNotElapsed(uint256 remainingCooldown);
 
 contract BondingCurve is ERC20 {
     // Define events
-    event TokenPurchased(address buyer, uint amount, uint totalPrice);
-    event TokenSold(address seller, uint amount, uint totalPrice);
+    event TokenPurchased(address buyer, uint256 amount, uint256 totalPrice);
+    event TokenSold(address seller, uint256 amount, uint256 totalPrice);
 
     // Define variables
-    uint public initialPrice;
-    uint public priceChangeRate;
+    uint256 public initialPrice;
+    uint256 public priceChangeRate;
     uint256 public reserveBalance;
     mapping(address => uint256) public lastBuyTimestamp;
     uint256 public constant cooldownPeriod = 1 days;
 
-    constructor(
-        uint256 _priceChangeRate,
-        uint256 _initialPrice
-    ) ERC20("Bonding Curve Continuous Token", "BCCT") {
+    constructor(uint256 _priceChangeRate, uint256 _initialPrice) ERC20("Bonding Curve Continuous Token", "BCCT") {
         initialPrice = _initialPrice;
         priceChangeRate = _priceChangeRate;
     }
@@ -49,10 +46,7 @@ contract BondingCurve is ERC20 {
         uint256 totalCost = getCostToBuy(purchaseAmount);
 
         if (msg.value < totalCost) {
-            revert InsufficientEther({
-                available: msg.value,
-                required: totalCost
-            });
+            revert InsufficientEther({available: msg.value, required: totalCost});
         }
         reserveBalance += totalCost;
 
@@ -63,14 +57,9 @@ contract BondingCurve is ERC20 {
 
         // Transfer any remaining ether to buyer
         if (msg.value - totalCost > 0) {
-            (bool sent, ) = payable(msg.sender).call{
-                value: msg.value - totalCost
-            }("");
+            (bool sent,) = payable(msg.sender).call{value: msg.value - totalCost}("");
             if (sent == false) {
-                revert InsufficientEther({
-                    available: msg.value - totalCost,
-                    required: 0
-                });
+                revert InsufficientEther({available: msg.value - totalCost, required: 0});
             }
         }
         // Emit event
@@ -81,26 +70,21 @@ contract BondingCurve is ERC20 {
      * @dev Sell tokens to receive ether.
      * @param depositAmount The amount of tokens to sell.
      */
-    function sellTokens(uint depositAmount) external {
+    function sellTokens(uint256 depositAmount) external {
         if (depositAmount == 0) {
             revert ZeroAmount();
         }
 
         // Check if cooldown period has elapsed since last buy
         if (block.timestamp < lastBuyTimestamp[msg.sender] + cooldownPeriod) {
-            revert CooldownNotElapsed(
-                block.timestamp - lastBuyTimestamp[msg.sender] + cooldownPeriod
-            );
+            revert CooldownNotElapsed(block.timestamp - lastBuyTimestamp[msg.sender] + cooldownPeriod);
         }
 
         // Calculate total price based on current pricePerToken and _amount
         uint256 totalCost = getCostToSell(depositAmount);
 
         if (totalCost > reserveBalance) {
-            revert InsufficientBalance({
-                available: reserveBalance,
-                required: totalCost
-            });
+            revert InsufficientBalance({available: reserveBalance, required: totalCost});
         }
 
         reserveBalance -= totalCost;
@@ -109,7 +93,7 @@ contract BondingCurve is ERC20 {
         _burn(msg.sender, depositAmount);
 
         // Transfer ether to seller
-        (bool sent, ) = payable(msg.sender).call{value: totalCost}("");
+        (bool sent,) = payable(msg.sender).call{value: totalCost}("");
         if (sent == false) {
             revert InsufficientEther({available: 0, required: totalCost});
         }
@@ -122,21 +106,13 @@ contract BondingCurve is ERC20 {
      * @param purchaseAmount The amount of tokens to purchase.
      * @return totalCost The total cost to purchase the specified amount of tokens.
      */
-    function getCostToBuy(
-        uint256 purchaseAmount
-    ) public view returns (uint256 totalCost) {
+    function getCostToBuy(uint256 purchaseAmount) public view returns (uint256 totalCost) {
         uint256 currentTotalSupply = totalSupply();
-        uint256 reserveBeforeTrade = (priceChangeRate *
-            currentTotalSupply ** 2) /
-            2e36 +
-            (initialPrice * currentTotalSupply) /
-            1e18;
+        uint256 reserveBeforeTrade =
+            (priceChangeRate * currentTotalSupply ** 2) / 2e36 + (initialPrice * currentTotalSupply) / 1e18;
         uint256 updatedTotalSupply = currentTotalSupply + purchaseAmount;
-        uint256 reserveAfterTrade = (priceChangeRate *
-            updatedTotalSupply ** 2) /
-            2e36 +
-            (initialPrice * updatedTotalSupply) /
-            1e18;
+        uint256 reserveAfterTrade =
+            (priceChangeRate * updatedTotalSupply ** 2) / 2e36 + (initialPrice * updatedTotalSupply) / 1e18;
         totalCost = reserveAfterTrade - reserveBeforeTrade;
         return totalCost;
     }
@@ -146,21 +122,13 @@ contract BondingCurve is ERC20 {
      * @param purchaseAmount The amount of tokens to sell.
      * @return totalCost The total proceeds from selling the specified amount of tokens.
      */
-    function getCostToSell(
-        uint256 purchaseAmount
-    ) public view returns (uint256 totalCost) {
+    function getCostToSell(uint256 purchaseAmount) public view returns (uint256 totalCost) {
         uint256 currentTotalSupply = totalSupply();
         uint256 updatedTotalSupply = currentTotalSupply - purchaseAmount;
-        uint256 reserveBeforeTrade = (priceChangeRate *
-            currentTotalSupply ** 2) /
-            2e36 +
-            (initialPrice * currentTotalSupply) /
-            1e18;
-        uint256 reserveAfterTrade = (priceChangeRate *
-            updatedTotalSupply ** 2) /
-            2e36 +
-            (initialPrice * updatedTotalSupply) /
-            1e18;
+        uint256 reserveBeforeTrade =
+            (priceChangeRate * currentTotalSupply ** 2) / 2e36 + (initialPrice * currentTotalSupply) / 1e18;
+        uint256 reserveAfterTrade =
+            (priceChangeRate * updatedTotalSupply ** 2) / 2e36 + (initialPrice * updatedTotalSupply) / 1e18;
         totalCost = reserveBeforeTrade - reserveAfterTrade;
         return totalCost;
     }
